+++ conf/default.php
@@ -5,5 +5,17 @@
  * @author Andreas Pazureck <andreas@pazureck.de>
  */
 
-//$conf['fixme']    = 'FIXME';
+$conf['chord_color'] = '#c94f2d';
+$conf['lyric_color'] = '#26312b';
+$conf['font_family'] = 'ui-monospace, "SFMono-Regular", Consolas, "Liberation Mono", monospace';
+$conf['lyrics_font_size'] = 1.0;
+$conf['chords_font_size'] = 1.0;
+$conf['line_spacing'] = 1.7;
+$conf['section_color'] = '#405347';
+$conf['section_background'] = '#fff2d7';
+$conf['section_spacing'] = 1.25;
+$conf['tooltip_behavior'] = 'hover_focus';
+$conf['section_style'] = 'accented';
+$conf['export_include_metadata'] = 1;
+$conf['export_font_family'] = 'Georgia, "Times New Roman", serif';